requirements.txt
python-dotenv
phidata
groq
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
sqlalchemy
pgvector
pypdf